fix invalid ca_bit
so the previous partial chain code effectively
set the CA bit on any certificate that was signed
by a certificate in the ca trust store.

we fix this by making sure the trusted_ca return
result is an actual ca certificate.

so erlang partial_chain requires the trusted_ca
you return to be identical to a cert already
in the chain. in an ideal world we would just
like to replace the chain and return a cert
that we consider a trusted ca. some servers
like s3.amazonaws.com return a root cert (but used
to be an intermediate cert?) in the chain and
erlang does not like this. so partial_chain is
meant to return this root cert as trusted_ca but
how is it possible to identify this as the root
cert? the cert in the trust store is not the same
as the cert the server gives. because the trust
store cert is self-signed and server cert is
signed by the CA that is no longer in the trust
store. so we just check if the public key is the
same. 

it is important to note that this creates
another security problem because a CA can serve up
a root cert with arbitrary extensions. for example
maybe a root cert has naming constraints but
by serving up a root cert without naming
constraints it is possible to subvert the naming
constraint policy. i'm not sure if erlang enforces
naming constraints on root cert so i'm not sure if
this introduces a vulnerability. also, there might
be problems with validity period and other
attributes on the certificate.