Try all backtrack branches and select best solution
When we decide which backtrack to take, backtrack current failed request
or backtrack its parents, we do not know which will lead to the best
solution. Instead of picking one we will try both and after resolution
is finished we will try to find the best solutions based on heuristics
such as using highest version number.

Resolution failures will have the same performance, but successful
resolutions with backtracks will be slower because they need to try more
backtracking branches.

Closes #571.