Fix resolver backtrack performance
This reverts a previous change that increased the number of backtracks we  do and the number of possible solutions we will have in the end. The change was made to get slightly better resolution results but introduced a performance issue so we are reverting this back until we have a better solution.