Add locked version to `mix hex.info <package>`
This PR adds the locked version of the package, if it is installed in
the current mix project.

If it is installed:

```bash
~/Development/hexpm [master] λ mix hex.info cowboy
Small, fast, modular HTTP server.

Config: {:cowboy, "~> 2.7"}
Locked version: 2.6.3
Releases: 2.7.0, 2.6.3, 2.6.2 (retired), 2.6.1, 2.6.0, 2.5.0, 2.4.0, 2.3.0, ...

Licenses: ISC
Links:
  GitHub: https://github.com/ninenines/cowboy
```

If it is *not* installed:
```bash
~/Development/hexpm [master] λ mix hex.info meetup
Elixir wrapper for the Meetup.com public API.

Config: {:meetup, "~> 0.1.2"}
Releases: 0.1.2, 0.1.1, 0.1.0

Licenses: GPL 3.0
Links:
  GitHub: https://github.com/supersimple/meetup
```

Thank you to @supersimple for pair-programming with me on this one!