Fix mix hex.outdated version sort assumption
When using mini_repo, this function is returning versions in a
different order than hex.outdated is expecting. To wit, one of our
applications is returning:

```elixir
["12.2.0", "12.1.5", "12.1.3", "12.0.0"]
```

This causes the `List.last/1` command here to return the incorrect
value.

I wasn't sure the best way to unit test this, but it's working for me
locally, so I figured I would open the PR as is and we could discuss
ways to test this behavior.