Add `--fetch-public-key FINGERPRINT` to `mix hex.repo add`
This allows Hex users to rely on the "/public_key" endpoint implemented
by the repositories to fetch the public key instead of doing it
manually.

The option requires the user to know the fingerprint of the key in order
to ensure more security. This prevents the key to be modified in
transit.