Unload hex application to fix compatability between Hex versions
Starting hex <0.21 from mix.exs crashes if hex 0.21+ is already loaded.
In the past we unloaded all current BEAM modules so that the development
version of Hex can be started when Hex is installed on your system.
We did not unload the application so when the application :mod key
changed we crashed. Fix by unloading and reloading the application as
well as the BEAM modules.