Add (yet unused) parser for .netrc files
This parser aims to implement the file format understood by cURL:

* Whitespace is not relevant and can be used freely to align/group
  values

* 'machine', 'login' and 'password' are supported, and must come on
  their own lines.

In particular, commented-out lines are not supported, nor is having all
keywords on a single line ('machine foo.com login yoyo password dyne').

Also, none of this was tested on a Windows system. It might be that line
endings or default file locations should be adjusted here.