Introduce Hex.Netrc.Cache agent
This agent is started at the very beginning; it avoids having to parse
.netrc files repeatedly. Instead of having to invoke
Hex.Netrc.Parser.parse/1, callers can use Hex.Netrc.Cache.parse/1 to
have the result of the parser get cached.

Both successful as well as unsuccessful parse results will be cached,
i.e. if parsing the given file failed once, that error will be returned
on subsequent cache requests for the same path. This is for two reasons:

* The cache process is rather short-lived: there's very little
  opportunity to fix any parse problems (e.g. by creating a .netrc file)
  while Hex is running.
* The cache receives multiple quiries in quick succession; afailing
  parse for a file which is stored on a high-latency file system (e.g.
  NFS) might slow down dependency resolution a lot.