Refactor hex.outdated
This unifies outdated dependency processing.

All behaviors should be the same, except for one edge case.

If there is only one dependency, the display will be the same as `single`. I'm not sure if that's an issue or not.

Assuming that the current UI is not meant to be programmatically accessed it should be fine.