Add client identifier header to HTTP requests (#1075)
* Add client identifier header to HTTP requests

Creates an anonymized repository identifier based on the SHA of the
first commit, hashed with SHA256 for additional privacy. The identifier
is sent as an `x-hex-client-id` header when available.

* Rename identifier header to x-hex-repo-id

* Respect HEX_REPO_IDENTIFIER environment variable

The new variable allows users to opt-out of identification by setting
the value to anything other than `1` or `true`. It remains enabled by
default.

* Add local caching to repo identifier lookup

Prevents fetching the same identifier on every client call by caching
the value in the current process dictionary. While the git command isn't
particularly slow, this should avoid spawning during repeated http
calls.

* Use Hex.State agent for identifier config

Switch from a raw `System.get_env` to using the normalized state
agent. Also prevent STDERR from leaking when called outside of a git
repository.

* Use Hex.State to store cached repo identifier

* Print error message on non-zero identifier message

* Synchronize cached repo identifier access

The repo identifier value is now cached with an agent, which also
serializes access to fetch the state. This also removes the value from
`Hex.State`, which was previously used for caching but is ultimately
unnecessary.

---------

Co-authored-by: Eric Meadows-Jönsson <eric.meadows.jonsson@gmail.com>